Refactor isoprobabilistic transforms
